Migration from Arduino IDE to Platformio
Created a project in Atom using Platformio to make the development easier.
