+++ velociruabtor_v2_teensy/.gitignore
@@ -1,0 +1,5 @@
+.pio
+.pioenvs
+.piolibdeps
+.clang_complete
+.gcc-flags.json
